dt-bindings: dp83867: add ti,led-function and ti,led-ctrl properties
With the TQ-Systems MBa7x (imx7-mba7.dtsi), a user of these properties
already sneaked in before they were properly specified. Add them to the
binding docs.

On top of the existing use (requiring to specify the raw register value
in the DTS), we propose a few convenience macros and defines.

Signed-off-by: Matthias Schiffer <matthias.schiffer@...tq-group.com>